% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalencetests-package.R
\docType{data}
\name{AMDA}
\alias{AMDA}
\title{Description of AMDA data}
\description{
Abbreviated Master Data File Final (May 16, 2006)-imputed.
}
\author{
Rob Cribbie \email{cribbie@yorku.ca}
}
\keyword{data}

