% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.1way.ww.R
\name{eq.1way.ww}
\alias{eq.1way.ww}
\title{Wellek-Welch equivalence test for k-groups}
\usage{
eq.1way.ww(dv, group, eps, alpha = 0.05)
}
\arguments{
\item{dv}{a numeric vector for the continous dependent variable}

\item{group}{grouping variable that can be coerced to a \code{factor}}

\item{eps}{numeric value defining the size of the equivalance interval}

\item{alpha}{the appropriate alpha level}
}
\description{
This function computes the Wellek-Welch equivalence test for k-groups. This addresses the research question about equivalence for means among k independent groups.
The Welch adjustment provides a robust correction for unequal vairances among the k groups.
}
\examples{
\dontrun{

dv <- rnorm(1000)
group <- rep(paste0('G',1:4), each=length(dv)/4)
eq.1way.ww(dv, group, eps = 0.5)

}
}
\author{
Rob Cribbie \email{cribbie@yorku.ca} and
  Phil Chalmers \email{rphilip.chalmers@gmail.com}
}

