% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.corr.R
\name{eq.corr}
\alias{eq.corr}
\alias{print.eq.corr}
\title{Test for lack of association: Equivalence-based correlation tests}
\usage{
eq.corr(dat, ei, alpha = 0.05, na.rm = TRUE, ...)

\method{print}{eq.corr}(x, ...)
}
\arguments{
\item{dat}{a data.frame or matrix containing only 2 variables}

\item{ei}{equivalence interval, in terms of the magnitude of a correlation}

\item{alpha}{desired alpha level}

\item{na.rm}{logical; remove missing values?}

\item{...}{additional arguments to be passed}

\item{x}{object of class \code{eq.corr}}
}
\value{
returns a \code{list} containing each analysis and their respective statistics
  and decision
}
\description{
Function performs a traditional Pearson correlation test, an equivalence based test of lack
of association, an equivalence based test of lack of association using a Fisher's z 
transformation, and an equivalence based test of lack of association with resampling.
}
\examples{
\dontrun{
#equivalence correlation test between v1 and v2 with an interval of .2
v1 <- rnorm(100)
v2 <- v1 + rnorm(100, 2)
dat <- data.frame(v1, v2)
plot(dat)
cor(dat)
eq.corr(dat, .2)
}
}
\author{
Rob Cribbie \email{cribbie@yorku.ca} and 
  Phil Chalmers \email{rphilip.chalmers@gmail.com}
}

