% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.hotT2.R
\name{eq.hotT2}
\alias{eq.hotT2}
\title{Hotelling T2}
\usage{
eq.hotT2(data, repeated, ei, alpha = 0.05)
}
\arguments{
\item{data}{a data.frame object}

\item{repeated}{a character vector of column names that are the dependent samples}

\item{ei}{equivalence interval, in the metric of Mahalanobis distance}

\item{alpha}{alpha level}
}
\description{
This function does the one sample Hotelling T2 analog for equivalence tests, determining if a sample's 
set of means on repeated measures are practically equivalent, given some pre-specified interval in the metric of Mahalanobis distance.
}
\examples{
k.obs <- 3
elements <- c(1, 0.7, 0.8, 1, 0.5,1)  

X <- diag(k.obs)  
X[lower.tri(X, diag = TRUE)] <- elements
X <- X + t(X) - diag(diag(X)) 
colnames(X) <- c('repA', 'repB', 'repC')
rownames(X) <- c('repA', 'repB', 'repC')

if(!require(mvtnorm)) install.packages(mvtnorm); library(mvtnorm) 
dat <- data.frame(rmvnorm(n=40, sigma=X))
names(dat) <- c('repA', 'repB', 'repC')
eq.hotT2(data=dat, repeated=c('repA', 'repB', 'repC'), ei=0.25)
}

