% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.tost.R
\name{eq.tost}
\alias{eq.tost}
\alias{x,}
\alias{y,}
\alias{ei,}
\alias{varequiv}
\alias{=}
\alias{FALSE,}
\alias{normality}
\alias{tr}
\alias{0.2,}
\alias{alpha}
\alias{0.05,}
\alias{na.rm}
\alias{TRUE,}
\alias{plot}
\alias{...}
\alias{print.eq.tost}
\title{Robust and Non-Robust Variants of the Two Independent Groups Equivalence Test 
using the two one-sided tests (TOST) methodology.}
\usage{
eq.tost(x, y, ei, varequiv = FALSE, normality = FALSE, tr = 0.2,
  alpha = 0.05, na.rm = TRUE, plot = TRUE, ...)

\method{print}{eq.tost}(x, ...)
}
\arguments{
\item{x}{a numeric vector for first sample}

\item{y}{a numeric vector for the second sample}

\item{ei}{numeric value defining the size (half-width) of the symmetric (around 0) equivalance interval}

\item{normality}{logical; If true, normality of x and y are assumed.}

\item{tr}{proportion of data to trim from each tail of each distribution (i.e., symmetric trimming). When \code{tr == 0}, the standard Schuirmann test
or Schuirmann-Welch is performed}

\item{alpha}{the maximum allowable Type I error rate}

\item{plot}{whether or not to print a graphic of the results}

\item{...}{additional arguments to be passed}

\item{varequal}{logical; If true, equal variances are assumed. Only applicable when tr == 0}

\item{x}{object of class \code{eq.tost}}
}
\value{
returns a \code{list}
}
\description{
This R function allows for the computation of the original Schuirmann TOST for 
the equivalence of two independent groups, the modified Schuirmann-Welch test 
(which does not require the variances to be equal), or the Schuirmann-Yuen test 
(if normality cannot be assumed), depending on the arguments specified. The original 
Schuirmann assumes equal variances and normality. The Schuirmann-Welch assumes 
normality but not equal variances. The Schuirmann-Yuen accounts for unequal variances 
and nonnormality using trimmed means and Winsorized variances.
}
\examples{
\dontrun{

x <- rnorm(100, 1)
y <- rnorm(100, 1)
# Original Schuirman TOST
eq.tost(x, y, ei=0.5, alpha = 0.05, varequiv=FALSE, normality=FALSE, plot=TRUE)
# Schuirmann-Welch
eq.tost(x, y, ei=0.5, alpha = 0.05, varequiv=FALSE, normality=TRUE, plot=TRUE)
# Schuirmann-Yuen TOST (default)
eq.tost(x, y, ei=0.5, alpha = 0.05, plot=TRUE)
#' # Schuirmann-Yuen TOST (10\% symmetric trimming)
eq.tost(x, y, ei=0.5, varequiv=FALSE, normality=FALSE, alpha = 0.05, print=TRUE) 
}
}
\references{
Schuirmann, D. J. (1987). A comparison of the two one-sided tests procedure and the power approach for assessing the equivalence of average bioavailability. \emph{Journal of pharmacokinetics and biopharmaceutics, 15(6)}, 657-680.
van Wieringen, K. & Cribbie, R. A. (2014). Robust normative comparison tests for evaluating clinical significance. \emph{British Journal of Mathematical and Statistical Psychology, 67}, 213-230.
Gruman, J., Cribbie, R. A., & Arpin-Cribbie, C. A. (2007). The effects of heteroscedasticity on tests of equivalence. \emph{Journal of Modern Applied Statistical Methods, 6}, 133-140.
}
\author{
Rob Cribbie \email{cribbie@yorku.ca}
}
