% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.hotT2.R
\name{equiv.hotT2}
\alias{equiv.hotT2}
\title{Hotelling T2
This function does the one sample Hotelling T2 analog for equivalence tests, determining if a sample's 
means on repeated measures are practically equivalent, given some pre-specified interval.}
\usage{
equiv.hotT2(data, repeated, ei, alpha = 0.05)
}
\arguments{
\item{data}{a data.frame object}

\item{repeated}{a character vector of column names that are the dependent samples}

\item{ei}{equivalence interval, in the metric of Mahalanobis distance}

\item{alpha}{alpha level}
}
\description{
Hotelling T2
This function does the one sample Hotelling T2 analog for equivalence tests, determining if a sample's 
means on repeated measures are practically equivalent, given some pre-specified interval.
}
\examples{
dat <- data.frame(pre=rnorm(6), post=rnorm(6,2), fu=rnorm(6,0.5), sex=c(rep('m', 2), rep('f', 4)) )
equiv.hotT2()
}

