% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivTrend.R
\name{equivTrend}
\alias{equivTrend}
\title{Trend test
This function takes as input the estimates obtained from a linear mixed model (fixed slope, random intercept model). 
This is essentially a TOST; instead of the mean difference as the parameter of interest, the slope is.}
\usage{
equivTrend(ei, slope, se, df, n, alpha = 0.05)
}
\arguments{
\item{ei}{equivalence interval}

\item{slope}{the estimated fixed effect slope obtained from some model}

\item{se}{the standard error of the slope}

\item{df}{the degrees of freedom obtained from the fitted model
equivTrend()}
}
\description{
Trend test
This function takes as input the estimates obtained from a linear mixed model (fixed slope, random intercept model). 
This is essentially a TOST; instead of the mean difference as the parameter of interest, the slope is.
}

