% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getContrast}
\alias{getContrast}
\title{create contrast matrices (for hotelling t2: adjacenet
mean differences; otherwise, pairwise)
pairwise_meanDiffs()}
\usage{
getContrast(k, type)
}
\description{
create contrast matrices (for hotelling t2: adjacenet
mean differences; otherwise, pairwise)
pairwise_meanDiffs()
}

