% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negTrend.R.R
\name{negTrend}
\alias{negTrend}
\alias{print.negTrend}
\title{Test of negligible trend (two one-sided tests approach)}
\usage{
negTrend(ei, slope, se, df, n, alpha = 0.05)

\method{print}{negTrend}(x, ...)
}
\arguments{
\item{ei}{equivalence interval}

\item{slope}{the estimated fixed effect slope obtained from some model}

\item{se}{the standard error of the slope}

\item{df}{the degrees of freedom obtained from the fitted model}

\item{x}{object of class \code{negTrend}}
}
\description{
This function takes as input the estimates obtained from a linear mixed model. Specifically, some fixed effect slope for time or some other ordered variable, its standard error, and the approximated degrees of freedom are used. 
This is essentially a TOST; instead of the mean difference as the parameter of interest, the slope is. 
The equivalence interval is then specified with the lower and upper limit bounds for what is considered to be minimum average slope deemed practically important. 
For analyses that include random slopes, one would used the fixed effect slope for the test of negligible trend.
One could also then evaluate the degree to which individual-level slope coefficients fall within negligible bounds, for assessing individual-level stability. 
Eg. What proportion of individual rates of (lack of) change had values that fell within the equivalence interval?
}

