% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pairwise_meanDiffs}
\alias{pairwise_meanDiffs}
\title{'Internal helper Function' for pairwise IUT based tests.}
\usage{
pairwise_meanDiffs(sample_means, allcontrasts)
}
\description{
This computes all pairwise mean differences.
}
\details{
pairwise_meanDiffs()
}

