% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw.std.R
\name{pw.std}
\alias{pw.std}
\title{Pairwise standardized}
\usage{
pw.std(data, repeated, ei, alpha = 0.05)
}
\arguments{
\item{data}{data set in data.frame format}

\item{repeated}{a character vector of column names designating the repeated measures}

\item{ei}{equivalence interval}

\item{alpha}{alpha level}
}
\description{
An equivalence test for repeated measures. Tests the null hypothesis that at least one pairwise mean difference 
is not practically equivalent, as defined by an equivalent interval.
}
\examples{
pw.std()
}

