% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{simRanIntSlope}
\alias{simRanIntSlope}
\title{'Internal helper Function' to simulate sample data
simRanIntSlope()}
\usage{
simRanIntSlope(sample_size, number_of_sample, equiv_interval, power)
}
\description{
'Internal helper Function' to simulate sample data
simRanIntSlope()
}

