% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.tost.CI.R
\name{eq.tost.CI}
\alias{eq.tost.CI}
\title{Independent samples equivalence test using confidence intervals
The two one-sided test for independent samples can be expressed using the confidence inclusion principle. If the confidence interval around a mean difference is within the equivalence bounds, then the mean difference is considered to be practically meaningless.}
\usage{
eq.tost.CI(x, y, ei, alpha = 0.05)
}
\arguments{
\item{ei}{equivalence interval}

\item{alpha}{desired alpha level}

\item{dat}{an N x 3 matrix or data.frame containing raw data used to compute the 
correlation matrix between variables. The input may also be a 1 x 3 vector of correlations
(r12, r13, and r23, respectively) and requires a sample size input (N)}

\item{n}{sample size when dat input is a vector of correlations}
}
\value{
returns a list containing the p-value, confidence interval, and statistical decision
}
\description{
Independent samples equivalence test using confidence intervals
The two one-sided test for independent samples can be expressed using the confidence inclusion principle. If the confidence interval around a mean difference is within the equivalence bounds, then the mean difference is considered to be practically meaningless.
}
\examples{
\dontrun{
#raw data
set.seed(1234)
x <- rnorm(100)
y <- rnorm(100)
ei=0.5
alpha=0.5
dat <- data.frame(dv=c(x,y),  group=c(rep('g1', length(x)), rep('g2', length(y)) ))
#Plot raw data distribution
ggplot(dat,aes(x=dv)) + 
  geom_histogram(data=subset(dat,group == 'g1'),fill = 'red', alpha = 0.2) +
  geom_histogram(data=subset(dat,group == 'g2'),fill = 'blue', alpha = 0.2)
equivInfo <- data.frame(name='mean difference', meanDiff=equivObj$meanDiff, lowCI=equivObj$lowCI, highCI=equivObj$hiCI)
require(ggplot2)
p <- ggplot(equivInfo, aes(x=name, y=meanDiff))+
  geom_pointrange(aes(ymin=lowCI, ymax=highCI)) +  xlab('comparison') +
  geom_hline(yintercept=c(-equivint, equivint), linetype='dashed', color='blue') + 
  coord_flip()
p
equivObj <- eq.TOST.CI(x,y, ei=0.50, alpha=.05)
}

}
\author{
Rob Cribbie \email{cribbie@yorku.ca} and 
  Phil Chalmers \email{rphilip.chalmers@gmail.com}
}

