% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.tost.R
\name{eq.tost}
\alias{...}
\alias{0.05,}
\alias{0.2,}
\alias{=}
\alias{FALSE,}
\alias{TRUE,}
\alias{alpha}
\alias{eq.tost}
\alias{equivint,}
\alias{na.rm}
\alias{normality}
\alias{print.eq.tost}
\alias{tost}
\alias{tr}
\alias{varequiv}
\alias{x,}
\alias{y,}
\title{Robust and Non-Robust Variants for the two one-sided equivalence tests (TOST): Test equivalence between 2 independent groups}
\usage{
eq.tost(x, y, equivint, varequiv = FALSE, normality = FALSE, tr = 0.2,
  alpha = 0.05, na.rm = TRUE, ...)

\method{print}{eq.tost}(x, ...)
}
\arguments{
\item{x}{a numeric vector for first sample}

\item{y}{a numeric vector for the second sample}

\item{equivint}{numeric value defining the size of the equivalance interval}

\item{normality}{logical; If true, normality of x and y are assumed.}

\item{tr}{proportion of data to trim. When \code{tr == 0}, equivintther the standard Schuirmann test
or Schuirmann-Welch is performed}

\item{alpha}{the appropriate alpha level}

\item{...}{additional arguments to be passed}

\item{varequal}{logical; If true, equal variances are assumed. Only applicable when tr == 0}

\item{x}{object of class \code{eq.tost}}
}
\value{
returns a \code{list}
}
\description{
This R function allows for the computation of the original Schuirmann two-sided t-test (TOST) for equivalence, the modified Schuirmann-Welch test (which does not require the variances to be equal), or the Schuirmann-Yuen test (if normality cannot be assumed), 
depending on the arguments specified. The original Schuirmann assumes equal variances and normality. The Schuirmann-Welch assumes normality but not equal variances. The Schuirmann-Yuen accounts for unequal variances and nonnormality using trimmed means and Winsorized variances.
}
\examples{
\dontrun{

x <- rnorm(100, 1)
y <- rnorm(100, 1)
# Original Schuirman TOST
eq.tost(x, y, equivint=0.5, alpha = 0.05, varequiv=FALSE, normality=FALSE)
# Schuirmann-Welch
eq.tost(x, y, equivint=0.5, alpha = 0.05, varequiv=FALSE, normality=TRUE)
# Schuirmann-Yuen TOST (default)
eq.tost(x, y, equivint=0.5, alpha = 0.05)
}
}
\author{
Rob Cribbie \email{cribbie@yorku.ca}
}
\references{
Schuirmann, D. J. (1987). A comparison of the two one-sided tests procedure and the power approach for assessing the equivalence of average bioavailability. \emph{Journal of pharmacokinetics and biopharmaceutics}, 15(6), 657-680.
}

