% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_rs.R
\name{equiv_rs}
\alias{equiv_rs}
\title{Test whether correlations or covariances are equivalent}
\usage{
equiv_rs(dat1, dat2, equiv_int, n1 = NULL, n2 = NULL, betas = FALSE,
  alpha = 0.05)
}
\arguments{
\item{dat1}{a matrix or data.frame containing raw data used to compute the first correlation/
covariance. A scalar input may be used as well specifying the sample correlation directly}

\item{dat2}{a matrix or data.frame containing raw data used to compute the second correlation/
covariance. A scalar input may be used as well specifying the sample correlation directly}

\item{equiv_int}{equivalence interval}

\item{n1}{sample size for first covariance set (required when input is a correlation)}

\item{n2}{sample size for second covariance set (required when input is a correlation)}

\item{betas}{logical; compare raw beta regression coefficients rather than correlations?}

\item{alpha}{desired alpha level}

\item{...}{additional arguments to be passed}
}
\value{
returns a list containing the coefficients, the equivalence interval, p-value, and
  statistical decision
}
\description{
Given raw data or known sample correlations, test whether correlation/covariance values
are equivalent within a specified interval. The null hypothesis is that the groups
are not equivalent in degree of association between two variables.
}
\details{
Function uses Anderson and Hauck's (1983) equivalence test. Because
the p-value derived from the test is only an approximation, the CIs may produce
results that fall outside of the equivalence interval at small sample sizes,
even with a statistically significant p-value. They therefore provide a measure
of precision, but cannot be used as a reject/fail to reject decision regarding
the null hypothesis.
}
\examples{
\dontrun{
#raw data
set.seed(1234)
dat1 <- cbind(rnorm(100), rnorm(100))
dat2 <- cbind(rnorm(200), rnorm(200))
equiv_rs(dat1, dat2, .2)
equiv_rs(dat1, dat2, .2, betas = TRUE)

par(mfrow=c(1,2))
plot(dat1)
plot(dat2)

#correlations input
r1 <- cor(dat1)[2,1]
r2 <- cor(dat2)[2,1]
equiv_rs(r1, r2, .2, n1 = 100, n2 = 200)

}
}
\author{
Rob Cribbie \email{cribbie@yorku.ca} and 
  Alyssa Counsell \email{counsela@yorku.ca}
}

