% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalencetests-package.R
\docType{package}
\name{equivalencetests-package}
\alias{equivalencetests}
\alias{equivalencetests-package}
\title{Equivalence tests}
\description{
Package for Equivalence Tests
}
\details{
The equivalencetests package provides statistical tools for equivalence testing in the null hypothesis testing framework. 
Available statistical tests include the following: 
ASSOCIATION-BASED
[eq.corr] One-sample lack of correlation/association; 
[equiv_rs] Two independent samples equivalence of correlations (equivalence-based independent samples) 
[equiv_drs] Dependent samples equivalence of correlations (equivalence-based dependent samples)

REPEATED-MEASURES / DEPENDENT SAMPLES
[eq.hotT2] k-dependent samples equivalence of repeated/multiple measures (multivariate approach) 
[pw.std] k-dependent samples equivalence of repeated/multiple measures (pairwise comparisons approach, using standardized metric)
[pw.unstd] k-dependent samples equivalence of repeated/multiple measures (pairwise comparisons approach, using unstand)
[equivTrend]

GROUP-BASED / INDEPENDENT-SAMPLES
[eq.tost] Two independent samples equivalence test of means
[eq.tost.CI] Two independent samples equivalence test of means using the confidence-interval inclusion approach
[eq.1way.ww] k independent samples equivalence tests of means
[eq.normcomp] normative comparison tests for evaluating clinical significance
}
\author{
Rob Cribbie \email{cribbie@yorku.ca}, Victoria Ng \email{ng.victoria.ky@gmail.com}, Alyssa Counsell \email{counsela@yorku.ca} and
  Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{package}

