\docType{package}
\name{equivalencetests-package}
\alias{equivalencetests}
\alias{equivalencetests-package}
\title{Equivalence tests}
\description{
Package for Equivalence Tests
}
\details{
Provides various functions for computing equivalence tests
in R.
}
\author{
Rob Cribbie \email{cribbie@yorku.ca} and Phil Chalmers
\email{rphilip.chalmers@gmail.com}
}
\keyword{package}

