% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{winvar}
\alias{winvar}
\title{Internal helper function to compute the gamma Winsorized variance for}
\usage{
winvar(x, tr = 0.2, na.rm = FALSE)
}
\description{
Internal helper function to compute the gamma Winsorized variance for
}

